Extract mapProductsTechnicalSpecs function (used by getProductsSpec middleware function) and it's helper functions with typings to separate module;
